Remove plans from the overall report file
