+++ tfcheck.py
@@ -19,18 +19,13 @@ parser.add_argument(
 )
 parser.add_argument(
     "-report",
-    help="print report to a file, removing plans if report length > 65336",
+    help="print detailed report to a file",
     type=argparse.FileType("a", encoding="UTF-8"),
 )
 parser.add_argument(
     "-hide-refresh",
     help="hide terraform state refresh output from report",
     action=argparse.BooleanOptionalAction,
-)
-parser.add_argument(
-    "-full-report",
-    help="print full report to a file",
-    type=argparse.FileType("a", encoding="UTF-8"),
 )
 
 env = Environment(loader=FileSystemLoader(f"{os.path.dirname(__file__)}/templates"))
@@ -140,19 +135,6 @@ def check(path: str, plan_args: Optional[str] = None) -> CheckResult:
     )
 
 
-def remove_plan(report: str) -> str:
-    report_lines = []
-    add_line = True
-    for line in report.splitlines():
-        if add_line and line == "<details><summary>Show Plan</summary>":
-            add_line = False
-        if add_line:
-            report_lines.append(line)
-        if not add_line and line == "</details>":
-            add_line = True
-    return "\n".join(report_lines)
-
-
 if __name__ == "__main__":
     args = parser.parse_args()
 
@@ -174,25 +156,18 @@ if __name__ == "__main__":
             )
         )
 
-    if args.report or args.full_report:
-        report = template.render(
-            path=result.path,
-            init_result=result.init_result(),
-            check_result=result.check_result_msg(),
-            fmt_result=result.fmt_result(),
-            validate_result=result.validate_result(),
-            plan_result=result.plan_result(),
-            plan_output=result.plan_output,
-            plan_msg=result.plan_msg(),
+    if args.report:
+        args.report.write(
+            template.render(
+                path=result.path,
+                init_result=result.init_result(),
+                check_result=result.check_result_msg(),
+                fmt_result=result.fmt_result(),
+                validate_result=result.validate_result(),
+                plan_result=result.plan_result(),
+                plan_output=result.plan_output,
+                plan_msg=result.plan_msg(),
+            )
         )
-
-        if args.full_report:
-            args.full_report.write(report)
-
-        if args.report:
-            if len(report) > 65536:
-                report = remove_plan(report)
-
-            args.report.write(report)
 
     sys.exit(result.exitcode())
